before switching to a single complex decision tree per person

+++ epiDNB/run.sh
@@ -9,12 +9,26 @@ fi
 echo
 echo "Triangulating..."
 rm -rf dbn.str.trifile
-gmtkTriangulate -strFile dbn.str -inputMasterFile dbn.master &&
+gmtkTriangulate -strFile dbn.str -inputMasterFile dbn.master
+
+if [ $? -eq 1 ]; then
+	exit 1
+fi
 
 echo
 echo "Decision tree processing..."
 rm -f dts/*.dts.index
-gmtkDTindex -inputMasterFile dbn.master && true
+gmtkDTindex -inputMasterFile dbn.master
 
-./train_cmd.sh data/dbn_train_100.txt &&
+if [ $? -eq 1 ]; then
+	exit 1
+fi
+
+./train_cmd.sh data/dbn_train_100.txt
+if [ $? -eq 1 ]; then
+	exit 1
+fi
 ./test_cmd.sh data/dbn_train_100.txt 
+if [ $? -eq 1 ]; then
+	exit 1
+fi
